data_gen/requirements.txt
# Core PDF generation
reportlab==4.0.9
PyPDF2==3.0.1

# Image manipulation for scan effects
Pillow==10.2.0
numpy==1.26.3
pdf2image==1.17.0

# Fake data generation
Faker==22.6.0

# Date handling
python-dateutil==2.8.2

# For adding realistic financial data
pandas==2.2.0

# Optional: For advanced PDF manipulation
pypdf==4.0.1

# Note: pdf2image requires poppler (system package)
# macOS: brew install poppler
# Ubuntu: sudo apt-get install poppler-utils
